+++ client/views/view_research.h
@@ -1,5 +1,5 @@
 /**************************************************************************
- Copyright (c) 1996-2020 Freeciv21 and Freeciv contributors. This file is
+ Copyright (c) 1996-2023 Freeciv21 and Freeciv contributors. This file is
  part of Freeciv21. Freeciv21 is free software: you can redistribute it
  and/or modify it under the terms of the GNU  General Public License  as
  published by the Free Software Foundation, either version 3 of the
@@ -13,7 +13,7 @@
 #include "fc_types.h"
 // client
 #include "repodlgs_g.h"
-#include "reqtree.h"
+#include "views/view_research_reqtree.h"
 
 class QComboBox;
 class QGridLayout;
