+++ client/views/view_research_reqtree.cpp
@@ -9,6 +9,12 @@
                   see https://www.gnu.org/licenses/.
  */
 
+/*
+ * This file contains functions to generate the GUI for the research
+ * requirements tree inside of the research view (formally known as the
+ * reqtree).
+ */
+
 // utility
 #include "log.h"
 
@@ -21,8 +27,8 @@
 // client
 #include "client_main.h"
 #include "options.h"
-#include "reqtree.h"
 #include "tileset/tilespec.h"
+#include "views/view_research_reqtree.h"
 
 #include "colors_g.h"
 #include "sprite_g.h"
